fix menus with more than 52 items
Pressing an accelerator will now toggle the next menu item with that
accelerator (or the previous item if no next item with that
accelerator exists).
In practice this only affects large menus, where the user must
position their cursor on the same line or above the item they would
like to toggle.